create ffmpeg class
trying to abstract out ffmpeg stuff. still need a way to edit various attribs such as level, breakaway, etc., from when original object is created
