+++ src/tests/unit/test_misc.py
@@ -123,33 +123,33 @@ def test_decide_whether_to_remove_rss_3(template_rss: TLShow):
 # ----- convert -----
 '''what else should we be testing here???'''
 
-def test_convert_1(template_rss: TLShow):
-    '''should convert and return name of file'''
-    assert type(template_rss.convert(input=download_test_file())) == str
+# def test_convert_1(template_rss: TLShow):
+#     '''should convert and return name of file'''
+#     assert type(template_rss.convert(input=download_test_file())) == str
 
-def test_convert_2(template_local: TLShow):
-    '''should convert and return name of file'''
-    assert template_local.convert(input=download_test_file()) == f'{template_local.show_filename}.wav'
-    os.remove(f'{template_local.show_filename}.wav') # actually converts the file so need to remove it
+# def test_convert_2(template_local: TLShow):
+#     '''should convert and return name of file'''
+#     assert template_local.convert(input=download_test_file()) == f'{template_local.show_filename}.wav'
+#     os.remove(f'{template_local.show_filename}.wav') # actually converts the file so need to remove it
 
-def test_convert_3(template_permalink: TLShow):
-    '''testing whether the breakaway functionality works as intended'''
-    template_permalink.breakaway = 30
-    test_file = template_permalink.convert(input=(download_test_file()))
-    duration = subprocess.getoutput(f"ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 {test_file}")
-    duration = float(duration)
-    duration = round(duration, 2)
-    assert duration == 30
-    os.remove(test_file)
+# def test_convert_3(template_permalink: TLShow):
+#     '''testing whether the breakaway functionality works as intended'''
+#     template_permalink.breakaway = 30
+#     test_file = template_permalink.convert(input=(download_test_file()))
+#     duration = subprocess.getoutput(f"ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 {test_file}")
+#     duration = float(duration)
+#     duration = round(duration, 2)
+#     assert duration == 30
+#     os.remove(test_file)
 
-def test_convert_4(template_permalink: TLShow):
-    '''assert an exception is raised when ffmpeg tries to convert a non-audio file'''
-    with open('test.mp3', mode='wb') as test_file:
-            a = requests.get('https://library.nashville.org/themes/custom/npl/logo.svg') #not an audio file
-            test_file.write(a.content)
-    with pytest.raises(Exception):
-        template_permalink.convert(input=test_file.name)
-    os.remove(test_file.name)
+# def test_convert_4(template_permalink: TLShow):
+#     '''assert an exception is raised when ffmpeg tries to convert a non-audio file'''
+#     with open('test.mp3', mode='wb') as test_file:
+#             a = requests.get('https://library.nashville.org/themes/custom/npl/logo.svg') #not an audio file
+#             test_file.write(a.content)
+#     with pytest.raises(Exception):
+#         template_permalink.convert(input=test_file.name)
+#     os.remove(test_file.name)
 
 # ----- check downloaded file -----
